user_service/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Werkzeug
PyJWT
Flask-Bcrypt
psycopg2-binary
Flask-Script
python-dotenv
requests
aiohttp
asyncio
pytest-mock
pytest-asyncio
werkzeug
pytz



